+++ samples/walletg.py
@@ -22,25 +22,13 @@ from pysui import PysuiConfiguration, SyncGqlClient
 # from pysui.sui.sui_pgql.pgql_clients import SuiGQLClient
 from pysui.sui.sui_constants import PYSUI_CLIENT_CONFIG_ENV
 
-from samples.cmd_argsg import build_parser
+from samples.cmd_argsg import build_parser, pre_config_pull
 from samples.cmdsg import SUI_CMD_DISPATCH
-
-
-def _group_pull(cli_arg: list, config: PysuiConfiguration) -> list:
-    """Check for group and set accordingly."""
-
-    if cli_arg.count("--group"):
-        ndx = cli_arg.index("--group")
-        nvl: str = cli_arg[ndx + 1]
-        config.make_active(group_name=nvl)
-
-    return cli_arg
 
 
 def main():
     """Entry point for demonstration."""
-    cfg = PysuiConfiguration(group_name=PysuiConfiguration.SUI_GQL_RPC_GROUP)
-    arg_line = _group_pull(sys.argv[1:].copy(), cfg)
+    cfg, arg_line = pre_config_pull(sys.argv[1:].copy())
     parsed = build_parser(arg_line, cfg)
     cmd_call = SUI_CMD_DISPATCH.get(parsed.subcommand, None)
     cfg.make_active(