+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.81.0] - Unpublished
+
+### Added
+
+### Fixed
+
+### Changed
+
+- [enhancement](https://github.com/FrankC01/pysui/issues/276) CLI utilities now accept `--config ` argument
+to use PysuiConfiguration from other than default path.
+
+### Removed
+
 ## [0.80.0] - 2025-02-24
 
 ### Added